harden broker honesty, gpu smoke, and module split
