+++ crates/codestory-contracts/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "codestory-contracts"
-version = "0.13.12"
+version = "0.14.0"
 edition = "2024"
 
 [dependencies]
