+++ crates/codestory-retrieval/src/lib.rs
@@ -42,8 +42,8 @@ pub use candidate::{CandidateHit, CandidateSource, RankFeatures};
 pub use candidate::{is_phantom_sidecar_hit, phantom_sidecar_candidates_only};
 pub use capabilities::SidecarCapabilities;
 pub use compose::{
-    BootstrapReport, DEFAULT_COMPOSE_REL_PATH, EmbedModelInventory, bootstrap_sidecars,
-    bootstrap_sidecars_with_profile, bootstrap_sidecars_with_runtime,
+    BootstrapReport, BootstrapSidecarsOptions, DEFAULT_COMPOSE_REL_PATH, EmbedModelInventory,
+    bootstrap_sidecars, bootstrap_sidecars_with_profile, bootstrap_sidecars_with_runtime,
     bootstrap_sidecars_with_runtime_progress, docker_available, embed_model_inventory,
     resolve_compose_file,
 };
@@ -55,11 +55,11 @@ pub use config::{
     default_sidecar_image_pins, embedding_server_launch_mode, sidecar_runtime_auto,
     sidecar_runtime_for_project, sidecar_runtime_for_project_with_run_id,
 };
-pub use embeddings::qdrant_vector_dim;
 pub use embeddings::{
-    BGE_BASE_EN_V1_5_GGUF, BGE_QUERY_PREFIX_DEFAULT, RETRIEVAL_EMBEDDING_DIM,
+    BGE_BASE_EN_V1_5_GGUF, BGE_QUERY_PREFIX_DEFAULT, EmbeddingRuntimeProbe, RETRIEVAL_EMBEDDING_DIM,
     embedding_backend_label, embedding_runtime_id, ensure_product_embedding_backend,
-    ensure_product_embedding_backend_for_runtime,
+    ensure_product_embedding_backend_for_runtime, probe_product_embedding_runtime,
+    qdrant_vector_dim,
 };
 pub use executor::{QueryExecutor, QueryResult, QueryTrace, StageTrace, cancellation_flag};
 pub use generation::{SIDECAR_SCHEMA_VERSION, SIDECAR_SEMANTIC_DOC_CONTRACT_CHANGED};
@@ -100,8 +100,8 @@ pub use scip_client::ScipClient;
 pub use sidecar::{
     NativeEmbeddingLaunchIdentityStatus, SidecarStateFile, ensure_native_embedding_launch_identity,
     native_embedding_launch_identity_status, sidecar_down, sidecar_down_for_project,
-    sidecar_down_for_runtime, sidecar_status, sidecar_up, sidecar_up_with_runtime,
-    sidecar_up_with_runtime_preserving_launch, strict_sidecar_status,
+    sidecar_down_for_runtime, sidecar_state_matches_runtime, sidecar_status, sidecar_up,
+    sidecar_up_with_runtime, sidecar_up_with_runtime_preserving_launch, strict_sidecar_status,
     strict_sidecar_status_for_profile, strict_sidecar_status_for_runtime,
 };
 pub use sidecar_search::{LiveSidecarSearch, SidecarSearch};
