+++ plugins/codestory/.github/plugin/plugin.json
@@ -1,7 +1,7 @@
 {
   "name": "codestory",
   "description": "CodeStory grounding for coding agents over the local codestory-cli runtime.",
-  "version": "0.13.12",
+  "version": "0.14.0",
   "author": {
     "name": "The Green Cedar",
     "url": "https://github.com/TheGreenCedar"
